API-862: Move Audit code in same bounded context
