+++ src/Akeneo/Apps/back/tests/Unit/spec/Domain/Audit/Model/Read/AppEventCountsSpec.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace spec\Akeneo\Apps\Audit\Domain\Model\Read;
+namespace spec\Akeneo\Apps\Domain\Audit\Model\Read;
 
-use Akeneo\Apps\Audit\Domain\Model\Read\DailyEventCount;
-use Akeneo\Apps\Audit\Domain\Model\Read\AppEventCounts;
+use Akeneo\Apps\Domain\Audit\Model\Read\AppEventCounts;
+use Akeneo\Apps\Domain\Audit\Model\Read\DailyEventCount;
 use PhpSpec\ObjectBehavior;
 
 /**
